Changed movement to use ApplyForce())

+++ Projectfiles/WeaponMagneticEffect.h
@@ -10,10 +10,10 @@
 
 @interface WeaponMagneticEffect : WeaponEffect <NSCopying>
 
-@property (nonatomic) int attractionPower; // Negative means repulsion
+@property (nonatomic) float attractionPower; // Negative means repulsion
 @property (nonatomic) float distanceAffected;
 
-- (id)initWithAttractionPower:(int) attraction withAffectedDistance:(float) distance;
+- (id)initWithAttractionPower:(float) attraction withAffectedDistance:(float) distance;
 - (id)copyWithZone:(NSZone *) zone;
 
 @end
